% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{in_rstudio}
\alias{in_rstudio}
\title{Determine whether base_path is open in RStudio}
\usage{
in_rstudio(base_path = proj_get())
}
\arguments{
\item{base_path}{The path to investigate}
}
\description{
This function is reproduced from the package 'usethis'.
For more details, see https://github.com/r-lib/usethis
}
