% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LA_puppies}
\alias{LA_puppies}
\title{Available puppies from the Los Angeles area extracted from Petfinder.com}
\format{A data frame with columns:
\describe{
\item{id}{Animal ID}
\item{organization_id}{Organization ID}
\item{url}{URL}
\item{type}{Type}
\item{species}{Species}
\item{breeds.primary}{Primary breed (if known)}
\item{breeds.secondary}{Secondary breed (if known)}
\item{breeds.mixed}{Is breed mixed? (TRUE if secondary breed exists)}
\item{breeds.unknown}{Is breed unknown?}
\item{age}{Age}
\item{gender}{Gender}
\item{size}{Size}
\item{attributes.spayed_neutered}{Is the animal spayed or neutered?}
\item{attributes.house_trained}{Is the animal house-trained?}
\item{attributes.special_needs}{Does the animal have special needs?}
\item{attributes.shots_current}{Are the animal's vaccinations current?}
\item{name}{Name}
\item{description}{Description of the animal's history, personality, and other characteristics}
\item{status}{Adoption status}
\item{published_at}{Date and time the animal's listing was posted}
\item{contact.email}{Contact email}
\item{contact.phone}{Contact phone}
\item{contact.address.address1}{Contact address: line 1}
\item{contact.address.city}{Contact address: city}
\item{contact.address.state}{Contact address: state}
\item{contact.address.postcode}{Contact address: postcode/zipcode}
\item{contact.address.country}{Contact address: country}
\item{colors.primary}{Primary color}
\item{colors.secondary}{Secondary color}
\item{coat}{Coat type}
\item{environment.children}{Can the animal live with children?}
\item{environment.dogs}{Can the animal live with dogs?}
\item{environment.cats}{Can the animal live with cats?}
\item{photos.small}{First photo, small size}
\item{photos.medium}{First photo, medium size}
\item{photos.large}{First photo, large size}
\item{photos.full}{First photo, full size}
\item{photos.small.1}{Second photo, small size}
\item{photos.medium.1}{Second photo, medium size}
\item{photos.large.1}{Second photo, large size}
\item{photos.full.1}{Second photo, full size}
\item{contact.address.address2}{Contact address: line 2}
\item{colors.tertiary}{Tertiary color}
}}
\source{
Petfinder API (V2); retrieved on April 25, 2019 \url{https://www.petfinder.com/developers/}
}
\usage{
LA_puppies
}
\description{
Available puppies from the Los Angeles area extracted from Petfinder.com
}
\examples{
\dontrun{
 LA_puppies
}
}
\keyword{datasets}
