% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NY_orgs}
\alias{NY_orgs}
\title{List of animal shelters in New York State extracted from Petfinder.com}
\format{A data frame with columns:
\describe{
\item{id}{Organization ID}
\item{name}{Organization name}
\item{email}{Organization email address}
\item{phone}{phone}
\item{address.city}{Organization address: city}
\item{address.state}{Organization address: state}
\item{address.postcode}{Organization address: postcode/zipcode}
\item{address.country}{Organization address: country}
\item{url}{URL}
\item{website}{Organization website}
\item{mission_statement}{Organization mission statement}
\item{adoption.policy}{Adoption policy}
\item{photos.small}{First photo, small size}
\item{photos.medium}{First photo, medium size}
\item{photos.large}{First photo, large size}
\item{photos.full}{First photo, full size}
\item{photos.small.1}{Second photo, small size}
\item{photos.medium.1}{Second photo, medium size}
\item{photos.large.1}{Second photo, large size}
\item{photos.full.1}{Second photo, full size}
\item{address.address1}{Organization address: line 1}
\item{address.address2}{Organization address: line 2}
\item{adoption.url}{Adoption URL}
\item{social_media.facebook}{Social media link: Facebook}
\item{social_media.youtube}{Social media link: YouTube}
\item{social_media.instagram}{Social media link: Instagram}
\item{social_media.twitter}{Social media link: Twitter}
\item{social_media.pinterest}{Social media link: Pinterest}
}}
\source{
Petfinder API (V2); retrieved on April 24, 2019 \url{https://www.petfinder.com/developers/}
}
\usage{
NY_orgs
}
\description{
List of animal shelters in New York State extracted from Petfinder.com
}
\examples{
\dontrun{
 NY_orgs
}
}
\keyword{datasets}
