% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorization.R
\name{petfindr_accesstoken}
\alias{petfindr_accesstoken}
\title{Generate an access token for the Petfinder API}
\usage{
petfindr_accesstoken(key, secret)
}
\arguments{
\item{key}{A key provided to the user by the Petfinder API}

\item{secret}{A secret provided to the user by the Petfinder API}
}
\description{
Generate an access token for the Petfinder API
}
\examples{
token <- petfindr_accesstoken(petfindr_key, petfindr_secret)
}
