% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petfindr_animal_search.R
\name{petfindr_animal_search}
\alias{petfindr_animal_search}
\title{Title}
\usage{
petfindr_animal_search(token, type = c("cat", "dog", "smallfurry",
  "barnyard", "bird", "horse", "reptile"), breed = NULL,
  size = c("small", "medium", "large", "xlarge"), gender = c("all",
  "male", "female", "unknown"), age = c("baby", "young", "adult",
  "senior"), color = NULL, coat = c("short", "medium", "long", "wire",
  "hairless", "curly"), status = c("adoptable", "adopted", "found"),
  name = NULL, organization = NULL, location = NULL,
  distance = NULL, sort = c("distance", "-distance", "name", "-name",
  "country", "-country", "state", "-state"), limit = 100, page = 1)
}
\arguments{
\item{token}{}

\item{type}{}

\item{breed}{}

\item{size}{}

\item{gender}{}

\item{age}{}

\item{color}{}

\item{coat}{}

\item{status}{}

\item{name}{}

\item{organization}{}

\item{location}{}

\item{distance}{}

\item{sort}{}

\item{limit}{}

\item{page}{}
}
\description{
Title
}
