% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petfindr_animaltypes.R
\name{petfindr_animaltypes}
\alias{petfindr_animaltypes}
\title{List all animal types found on Petfinder}
\usage{
petfindr_animaltypes(token, type = NULL)
}
\arguments{
\item{token}{An access token}

\item{type}{One of the eight available types: "dog", "cat", "rabbit", "small & furry", "horse", "bird", "scales, fins, & other", or "barnyard". If no type is provided, all types are returned.}
}
\value{
A tibble listing the desired animal types with their available coats, colors, and genders
}
\description{
Longer description of what the function does
}
\examples{
petfindr_animaltypes(token)
petfindr_animaltypes(token, "dog")
petfindr_animaltypes(token)$name
}
