% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorization.R
\name{petfindr_save_credentials}
\alias{petfindr_save_credentials}
\title{Save PetFindr key and secret to .Rprofile}
\usage{
petfindr_save_credentials(key = NULL, secret = NULL,
  scope = c("project", "user"))
}
\arguments{
\item{key}{A key provided to the user by the Petfinder API}

\item{secret}{A secret provided to the user by the Petfinder API}

\item{scope}{The scope of the .Rprofile to which the key and secret should be saved; one of "project" or "user".}
}
\value{
None
}
\description{
This function allows the user to save their Petfinder API key and/or secret to their user or project .Rprofile
}
\examples{
petfindr_save_credentials(petfindr_key, petfindr_secret)
}
