% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorization.R
\name{petfindr_setup}
\alias{petfindr_setup}
\title{Get started with PetFindr}
\usage{
petfindr_setup()
}
\value{
None
}
\description{
Running this function will open the website for Petfinder's API.
}
\examples{
petfindr_setup()
}
