% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorization.R
\name{pf_accesstoken}
\alias{pf_accesstoken}
\title{Generate an access token for the Petfinder API}
\usage{
pf_accesstoken(key, secret)
}
\arguments{
\item{key}{A key provided to the user by the Petfinder API}

\item{secret}{A secret provided to the user by the Petfinder API}
}
\value{
An access token for the PetFindr API
}
\description{
Generate an access token for the Petfinder API
}
\examples{
token <- pf_accesstoken(petfindr_key, petfindr_secret)
}
