% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pf_breeds}
\alias{pf_breeds}
\title{Available breeds by type from Petfinder.com}
\format{A data frame with columns:
\describe{
\item{name}{Available animal types}
\item{breeds}{Available breeds for each animal type, separated by commas}
}}
\source{
Petfinder API (V2); retrieved on April 23, 2019 \url{https://www.petfinder.com/developers/}
}
\usage{
pf_breeds
}
\description{
Available breeds by type from Petfinder.com
}
\examples{
\dontrun{
 pf_breeds
}
}
\keyword{datasets}
