% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_build_query.R
\name{pf_build_query}
\alias{pf_build_query}
\title{Build a query string}
\usage{
pf_build_query(args = NULL)
}
\arguments{
\item{args}{The list of arguments from the parent function}
}
\value{
A query string
}
\description{
Build a query string
}
