% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_find_organizations.R
\name{pf_find_organizations}
\alias{pf_find_organizations}
\title{find_organizations}
\usage{
pf_find_organizations(token, name = NULL, location = NULL,
  distance = NULL, state = NULL, country = NULL, sort = NULL,
  limit = 100, page = 1)
}
\arguments{
\item{token}{An access token}

\item{name}{Names of organizations}
}
\value{
A dataframe listing information of the desired organizations
}
\description{
Longer description of what the function does
}
\examples{
organizations_of_interest <- pf_find_organizations(token, country = "US", limit = 100, page = 1, sort = "state")
}
