% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_find_orgid.R
\name{pf_find_orgid}
\alias{pf_find_orgid}
\title{Obtain a data frame of organization id using pf_find_pets}
\usage{
pf_find_orgid(data)
}
\arguments{
\item{data}{a data frame obtained from pf_find_pets}
}
\value{
(a data frame obtained by organization id from pf_find_pets)
}
\description{
Retrieve a data frame of information about organizations using the organization id information
from pf_find_pets.
}
\examples{
\dontrun{
pf_dat <- pf_find_pets(token=tokne, location=50014, type="dog")
pf_find_orgid(pf_dat)
}
}
