% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_list_breeds.R
\name{pf_list_breeds}
\alias{pf_list_breeds}
\title{List available breeds for a given animal type}
\usage{
pf_list_breeds(token, type = NULL)
}
\arguments{
\item{token}{An access token, provided by \code{\link{pf_accesstoken}}.}

\item{type}{One of the available animal types for which to return breed information. A full list of animal types can be found by running \code{\link{pf_list_types}}.}
}
\value{
A character vector containing the breed names for the specified animal type.
}
\description{
This function allows the user to specify an animal type found on Petfinder.com and returns a vector containing all available breeds for that animal type.
}
\examples{
\dontrun{
pf_list_breeds(token, type = "dog")
pf_list_breeds(token, type = "barnyard")
}
}
