% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_map_locations.R
\name{pf_locate_organizations}
\alias{pf_locate_organizations}
\title{Get location information for searched pets}
\usage{
pf_locate_organizations(token, animal_df)
}
\arguments{
\item{token}{An access token, provided by \code{\link{pf_accesstoken}}.}

\item{animal_df}{A data frame of animal information output from \code{\link{pf_find_pets}}.}
}
\value{
The original data frame supplemented with more detailed organization location information.
}
\description{
Get location information for searched pets
}
