% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_map_organizations.R
\name{pf_map_organizations}
\alias{pf_map_organizations}
\title{NAME/TITLE OF FUNCTION}
\usage{
pf_map_organizations(id)
}
\arguments{
\item{id}{}
}
\value{
(Fill this in)
}
\description{
NAME/TITLE OF FUNCTION
}
\examples{
pf_animal_search(token, type = "dog")
animal_dat <- pf_find_pets(token, location=50014, distance=10, type="dog", page="all")
id_ex <- unique(animal_dat$organization_id)
pf_organization_map(id=id_ex)
}
