% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_map_animals.R
\name{pf_merge_organizations}
\alias{pf_merge_organizations}
\title{Merge Organization Data to Animal Data}
\usage{
pf_merge_organizations(token, animal_df)
}
\arguments{
\item{token}{An access token, provided by pf_accesstoken(key, secret).}

\item{animal_df}{A data frame of animal information output from pf_find_pets().}
}
\value{
The original data frame supplemented with more detailed organization information
}
\description{
Merge Organization Data to Animal Data
}
