% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_petfind.R
\name{pf_petfind}
\alias{pf_petfind}
\title{Better pet finding--in progress}
\usage{
pf_petfind(token, page = 1, limit = 20, sort = c("recent", "-recent",
  "distance", "-distance"), ...)
}
\arguments{
\item{token}{}

\item{limit}{}

\item{sort}{}

\item{...}{}
}
\value{
a df (ideally)
}
\description{
Better pet finding--in progress
}
\examples{
pf_petfind(token, gender = "female", age = "baby", type = "dog")
}
