% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_viewphotos.R
\name{pf_photo_view}
\alias{pf_photo_view}
\title{Title slideshow all available photos of the animals of interest in R}
\usage{
pf_photo_view(search_result, size = c("small", "medium", "large",
  "full"))
}
\arguments{
\item{search_result}{}

\item{size}{}
}
\description{
Title slideshow all available photos of the animals of interest in R
}
\examples{
animals_of_interest<-pf_animal_search(token, location = 50014, distance = 100, type = "dog",breed = "corgi", gender = c("male", "female"),
limit=100, page="all", sort = "distance")
pf_photo_view(search_result= animals_of_interest, size= "medium")
}
