% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_run_Shiny.R
\name{pf_run_Shiny}
\alias{pf_run_Shiny}
\title{Run PetFindr Shiny app}
\usage{
pf_run_Shiny()
}
\description{
Launch the PetFindr Shiny application that displays animal location information.
}
\examples{
\dontrun{
pf_run_Shiny()
}
}
