% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_run_Shiny.R
\name{pf_runShinyApp}
\alias{pf_runShinyApp}
\title{Run petorg example}
\usage{
pf_runShinyApp()
}
\description{
Launch a Shiny app that shows a demo of what can be done with
}
\details{
\code{PetFindr::petorg}.
}
\examples{
\dontrun{
pf_ShinyApp()
}
}
