% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorization.R
\name{pf_save_credentials}
\alias{pf_save_credentials}
\title{Save Petfinder API (V2) key and secret to .Rprofile}
\usage{
pf_save_credentials(key = NULL, secret = NULL)
}
\arguments{
\item{key}{A key provided to the user by the Petfinder API (V2).}

\item{secret}{A secret provided to the user by the Petfinder API (V2).}
}
\value{
None
}
\description{
This function allows the user to save their Petfinder API (V2) key and/or secret to their .Rprofile for future use.
}
\examples{
\dontrun{
pf_save_credentials(key, secret)
}
}
