% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorization.R
\name{pf_setup}
\alias{pf_setup}
\title{Get started with PetFindr}
\usage{
pf_setup()
}
\value{
None
}
\description{
Running this function will open a browser to "https://www.petfinder.com/developers/" and prompt the user to register for Petfinder's API (V2).
}
\examples{
\dontrun{
pf_setup()
}
}
