% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pf_types}
\alias{pf_types}
\title{Available animal types from Petfinder.com, with their respective coat, color, and gender options}
\format{A data frame with columns:
\describe{
\item{name}{Available animal types}
\item{coats}{Available coats for each animal type; if multiple, values are separated by commas}
\item{colors}{Available colors for each animal type; if multiple, values are separated by commas}
\item{genders}{Available coats for each animal type; if multiple, values are separated by commas}
}}
\source{
Petfinder API (V2); retrieved on April 23, 2019 \url{https://www.petfinder.com/developers/}
}
\usage{
pf_types
}
\description{
Available animal types from Petfinder.com, with their respective coat, color, and gender options
}
\examples{
\dontrun{
 pf_types
}
}
\keyword{datasets}
