% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_view_photos.R
\name{pf_view_photos}
\alias{pf_view_photos}
\title{Slideshow all available photos of the animals of interest in R}
\usage{
pf_view_photos(animal_df, size = c("small", "medium", "large", "full"))
}
\arguments{
\item{animal_df}{A data frame of the animals of interest that have been searched for (using pf_find_pet)}

\item{size}{The favorite size of the animal photos to be shown}
}
\value{
A slideshow of the searched pets
}
\description{
Slideshow all available photos of the animals of interest in R
}
\examples{
\dontrun{
  puppies <- pf_find_pets(token, type = "dog", age = "baby", breed= "corgi" )
  pf_view_photos(animal_df = puppies, size = "small")
  
  bunnies <- pf_find_pets(token = token, type = "rabbit", age = "baby", page= 1:2)
  pf_view_photos(animal_df = bunnies, size = "full")
}
}
