% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restart_rstudio}
\alias{restart_rstudio}
\title{Restart RStudio.}
\usage{
restart_rstudio(message = NULL)
}
\description{
This function is reproduced from the package 'usethis'.
For more details, see https://github.com/r-lib/usethis
}
