% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run petorg example}
\usage{
runExample()
}
\description{
Launch a Shiny app that shows a demo of what can be done with
}
\details{
\code{PetFindr::petorg}.
}
\examples{
## Only run this example in interactive R sessions
runExample()


}
