% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorization.R
\name{pf_accesstoken}
\alias{pf_accesstoken}
\title{Generate an access token for the Petfinder API (V2)}
\usage{
pf_accesstoken(key = NULL, secret = NULL)
}
\arguments{
\item{key}{A key provided to the user by the Petfinder API (V2)}

\item{secret}{A secret provided to the user by the Petfinder API (V2)}
}
\value{
Create a object called "token" which has the value of access token for the Petfinder API (V2)
}
\description{
Generate an access token for the Petfinder API (V2)
}
\examples{
\dontrun{
pf_accesstoken(petfindr_key, petfindr_secret)
}
}
