% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_breeds.R
\name{pf_breeds}
\alias{pf_breeds}
\title{List available breeds for a specific animal type on Petfinder}
\usage{
pf_breeds(token, type = c("dog", "cat", "rabbit", "small & furry",
  "horse", "bird", "scales, fins, & other", "barnyard"))
}
\arguments{
\item{token}{An access token for the Petfinder API}

\item{type}{One of the eight available types. If no type is specified, dog breeds are returned.}
}
\value{
A character vector containing the breed names for the given animal type
}
\description{
This function takes in one of the eight animal types found on Petfinder.com and returns a vector containing all available breeds for that animal type
}
\examples{
pf_breeds(token, type = "dog")
pf_breeds(token, type = "cat")

}
