% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_animal_search.R, R/pf_find_pets.R
\name{pf_find_pets}
\alias{pf_find_pets}
\title{Title}
\usage{
pf_find_pets(token = NULL, type = NULL, breed = NULL, size = NULL,
  gender = NULL, age = NULL, color = NULL, coat = NULL,
  status = NULL, name = NULL, organization = NULL, location = NULL,
  distance = NULL, sort = "recent", page = 1, limit = 20)

pf_find_pets(token = NULL, type = NULL, breed = NULL, size = NULL,
  gender = NULL, age = NULL, color = NULL, coat = NULL,
  status = NULL, name = NULL, organization = NULL, location = NULL,
  distance = NULL, sort = "recent", page = 1, limit = 20)
}
\arguments{
\item{token}{}

\item{type}{}

\item{breed}{}

\item{size}{}

\item{gender}{}

\item{age}{}

\item{color}{}

\item{coat}{}

\item{status}{}

\item{name}{}

\item{organization}{}

\item{location}{}

\item{distance}{}

\item{sort}{}

\item{page}{}

\item{limit}{}

\item{token}{An access token, provided by pf_accesstoken(key, secret).}

\item{type}{The type(s) of animals to be found. A full list of animal types, along with their respective coat and color options, can be found by running pf_animaltypes(token).}

\item{breed}{The breed(s) of animals to be found. A full list of breeds for a given animal type can be found by running pf_breeds(token, type).}

\item{size}{The size(s) of animals to be found. Possible values are "small", "medium", "large", and "xlarge".}

\item{gender}{The gender(s) of animals to be found. Possible values are "male", "female", and "unknown".}

\item{age}{The age(s) of animals to be found. Possible values are "baby", "young", "adult", and "senior".}

\item{color}{The color(s) of animals to be found. A full list of animal types, along with their respective coat and color options, can be found by running pf_animaltypes(token).}

\item{coat}{The coat(s) of animals to be found. A full list of animal types, along with their respective coat and color options, can be found by running pf_animaltypes(token).}

\item{status}{The status of animals to be found. Possible values are "adoptable", "adopted", or "found".}

\item{name}{The name of animals to be found (includes partial matches; e.g. "Fred" will return "Alfredo" and "Frederick").}

\item{organization}{The organization(s) associated with animals to be found. Values should be provided as identification numbers.}

\item{location}{The location of animals to be found. Values can be specified as "\link{City}, \link{State}", "\link{latitude}, \link{longitude}", or "\link{postal code}".}

\item{distance}{The distance, in miles, from the provided location to find animals. Note that location is required to use distance.}

\item{sort}{The attribute on which to sort results. Possible attributes are "recent", "-recent", "distance", or "-distance".}

\item{page}{The page of results to return. Default is 1.}

\item{limit}{The maximum number of results to return per page. Default is 20, maximum is 100.}
}
\value{
(Fill this in)

A data frame of results matching the search parameters
}
\description{
Title

Search for pets from Petfinder.com
}
\examples{
pf_find_pets(token, type = "dog")
animals_of_interest <- pf_find_pets(token, location = 50014, distance = 150, type = "dog", breed = "corgi", gender = c("male", "female"), age = "baby", limit=100, page="all", sort ="distance")
puppies <- pf_find_pets(token, type = "dog", age = "baby", gender = "blat")
}
