% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_find_pets.R
\name{pf_find_pets}
\alias{pf_find_pets}
\title{Find Pets}
\usage{
pf_find_pets(token = NULL, type = NULL, breed = NULL, size = NULL,
  gender = NULL, age = NULL, color = NULL, coat = NULL,
  status = NULL, name = NULL, organization = NULL, location = NULL,
  distance = NULL, sort = "recent", page = 1, limit = 20)
}
\arguments{
\item{token}{An access token, provided by pf_accesstoken(key, secret).}

\item{type}{The type(s) of animals to be found. A full list of animal types, along with their respective coat and color options, can be found by running pf_list_types(token).}

\item{breed}{The breed(s) of animals to be found. A full list of breeds for a given animal type can be found by running pf_breeds(token, type).}

\item{size}{The size(s) of animals to be found. Possible values are "small", "medium", "large", and "xlarge".}

\item{gender}{The gender(s) of animals to be found. Possible values are "male", "female", and "unknown".}

\item{age}{The age(s) of animals to be found. Possible values are "baby", "young", "adult", and "senior".}

\item{color}{The color(s) of animals to be found. A full list of animal types, along with their respective coat and color options, can be found by running pf_list_types(token).}

\item{coat}{The coat(s) of animals to be found. A full list of animal types, along with their respective coat and color options, can be found by running pf_list_types(token).}

\item{status}{The status of animals to be found. Possible values are "adoptable", "adopted", or "found".}

\item{name}{The name of animals to be found (includes partial matches; e.g. "Fred" will return "Alfredo" and "Frederick").}

\item{organization}{The organization(s) associated with animals to be found. Values should be provided as identification numbers.}

\item{location}{The location of animals to be found. Values can be specified as "<City>, <State>", "<latitude>, <longitude>", or "<postal code>".}

\item{distance}{The distance, in miles, from the provided location to find animals. Note that location is required to use distance.}

\item{sort}{The attribute on which to sort results. Possible attributes are "recent", "-recent", "distance", or "-distance".}

\item{page}{The page(s) of results to return (default = 1).}

\item{limit}{The maximum number of results to return per page (maximum = 100).}
}
\value{
A data frame of results matching the search parameters
}
\description{
Retrieve a data frame of information about pets that are listed on
Petfinder.com via the Petfinder API (V2). Filter searches based on
characteristics such as animal type, breed, size, age, or location.
}
\examples{
\dontrun{
puppies <- pf_find_pets(token, type = "dog", age = "baby", page = 1:5)
}
}
