% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_list_types.R
\name{pf_list_types}
\alias{pf_list_types}
\title{List animal types with their respective coat, color, and gender options}
\usage{
pf_list_types(token)
}
\arguments{
\item{token}{An access token}
}
\value{
A tibble listing all available animal types with their respective coat, color, and gender options.
}
\description{
This function returns the available animal types from Petfinder.com, along with each type's available coat, color, and gender options.
}
\examples{
\dontrun{
pf_list_types(token)
pf_list_types(token)$name
}
}
