% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_map_organizations.R
\name{pf_map_organizations}
\alias{pf_map_organizations}
\title{Drawing a map and get data frame with organization id.}
\usage{
pf_map_organizations(id)
}
\arguments{
\item{id}{Organization id}
}
\value{
A data frame using organzation and print a leaflet map
}
\description{
Drawing a map and get data frame with organization id.
}
\examples{
\dontrun{

animal_dat <- pf_find_pets(token, location=50014, distance=10, type="dog")
id_ex <- unique(animal_dat$organization_id)
pf_map_organizations(id=id_ex)
}
}
