% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_view_photos.R
\name{pf_photo_view}
\alias{pf_photo_view}
\title{Title slideshow all available photos of the animals of interest in R}
\usage{
pf_photo_view(search_result, size = c("small", "medium", "large",
  "full"))
}
\arguments{
\item{search_result}{}

\item{size}{}
}
\value{
SOMETHING MUST GO HERE
}
\description{
Title slideshow all available photos of the animals of interest in R
}
