% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorization.R
\name{pf_setup}
\alias{pf_setup}
\title{Get started with PetFindr}
\usage{
pf_setup()
}
\value{
None
}
\description{
Running this function will open the website for Petfinder's API (V2) and prompt the user to sign up for an account.
}
\examples{
pf_setup()
}
