% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_view_photos.R
\name{pf_view_photos}
\alias{pf_view_photos}
\title{Slideshow all available photos of the animals of interest in R}
\usage{
pf_view_photos(animal_df, size = c("small", "medium", "large", "full"))
}
\arguments{
\item{animal_df}{YOU NEED TO DESCRIBE THIS}

\item{size}{YOU NEED TO DESCRIBE THIS}
}
\value{
A slideshow of the searched pets
}
\description{
Slideshow all available photos of the animals of interest in R
}
\examples{
\dontrun{
corgis <- pf_find_pets(token, location = 50014, distance = 100, type = "dog",
    breed = "corgi", gender = c("male", "female"), limit = 100, page = "all", 
    sort = "distance")
pf_view_photos(animal_df = corgis, size = "medium")
}
}
