% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_view_photos.R
\name{pf_view_photos}
\alias{pf_view_photos}
\title{View photos of pets}
\usage{
pf_view_photos(animal_df, size = c("small", "medium", "large", "full"))
}
\arguments{
\item{animal_df}{A data frame of animal search results from \code{\link{pf_find_pets}}.}

\item{size}{The desired size of the animal photos to be shown. One of "small", "medium", "large", or "full".}
}
\value{
A slideshow of animal pictures
}
\description{
View the photos of searched pets in a slideshow format.
}
\examples{
\dontrun{
  puppies <- pf_find_pets(token, type = "dog", age = "baby", breed = "corgi")
  pf_view_photos(animal_df = puppies, size = "small")
  
  bunnies <- pf_find_pets(token, type = "rabbit", age = "baby", limit = 10)
  pf_view_photos(animal_df = bunnies, size = "full")
  
  birds<- pf_find_pets(token, type= "Bird")
  pf_view_photos(birds)
}
}
